Reverted to [9025] because commit [9026] failed (it was incomplete with a broken working dir).  - Legacy-Id: 9027 Note: SVN reference [9026] has been migrated to Git commit 4a3749a66b69fe8663804800a22a9198269620a0

+++ django/contrib/gis/gdal/srs.py
@@ -1,5 +1,5 @@
 """
-  The Spatial Reference class, represents OGR Spatial Reference objects.
+  The Spatial Reference class, represensents OGR Spatial Reference objects.
 
   Example:
   >>> from django.contrib.gis.gdal import SpatialReference
@@ -97,8 +97,7 @@ class SpatialReference(GDALBase):
 
     def __del__(self):
         "Destroys this spatial reference."
-        if self._ptr:
-            capi.release_srs(self._ptr)
+        if self._ptr: capi.release_srs(self._ptr)
 
     def __getitem__(self, target):
         """
@@ -106,7 +105,7 @@ class SpatialReference(GDALBase):
         doesn't exist.  Can also take a tuple as a parameter, (target, child),
         where child is the index of the attribute in the WKT.  For example:
 
-        >>> wkt = 'GEOGCS["WGS 84", DATUM["WGS_1984, ... AUTHORITY["EPSG","4326"]]'
+        >>> wkt = 'GEOGCS["WGS 84", DATUM["WGS_1984, ... AUTHORITY["EPSG","4326"]]')
         >>> srs = SpatialReference(wkt) # could also use 'WGS84', or 4326
         >>> print(srs['GEOGCS'])
         WGS 84
@@ -120,7 +119,7 @@ class SpatialReference(GDALBase):
         0
         >>> print(srs['UNIT|AUTHORITY']) # For the units authority, have to use the pipe symbole.
         EPSG
-        >>> print(srs['UNIT|AUTHORITY', 1]) # The authority value for the units
+        >>> print(srs['UNIT|AUTHORITY', 1]) # The authority value for the untis
         9122
         """
         if isinstance(target, tuple):
@@ -177,14 +176,10 @@ class SpatialReference(GDALBase):
     @property
     def name(self):
         "Returns the name of this Spatial Reference."
-        if self.projected:
-            return self.attr_value('PROJCS')
-        elif self.geographic:
-            return self.attr_value('GEOGCS')
-        elif self.local:
-            return self.attr_value('LOCAL_CS')
-        else:
-            return None
+        if self.projected: return self.attr_value('PROJCS')
+        elif self.geographic: return self.attr_value('GEOGCS')
+        elif self.local: return self.attr_value('LOCAL_CS')
+        else: return None
 
     @property
     def srid(self):
@@ -328,7 +323,6 @@ class SpatialReference(GDALBase):
         "Returns the XML representation of this Spatial Reference."
         return capi.to_xml(self.ptr, byref(c_char_p()), dialect)
 
-
 class CoordTransform(GDALBase):
     "The coordinate system transformation object."
 
@@ -342,8 +336,7 @@ class CoordTransform(GDALBase):
 
     def __del__(self):
         "Deletes this Coordinate Transformation object."
-        if self._ptr:
-            capi.destroy_ct(self._ptr)
+        if self._ptr: capi.destroy_ct(self._ptr)
 
     def __str__(self):
         return 'Transform from "%s" to "%s"' % (self._srs1_name, self._srs2_name)
